+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- GetContentProtections methods to AdaptationSet and Representation
+- GetContentProtections, GetMimeType, and GetSegmentTemplate methods for AdaptationSet and Representation
 
 ## [0.10.0] - 2023-05-26